Fixes to installation docs; remove unused "login"
Add a notes to INSTALLATION.md explaining:
  - What to set the callback URL
  - how "UNHANGOUT_ADMIN_EMAILS" works

Also remove unused "login" route and method from
lib/unhangout-server.js.  Passport doesn't seem to use "failureRedirect"
when it has an explicit handler method set -- but even if it did, the
login route was calling an undefined view.  Currently, a auth failure
from "google" results in a 500 with error message, which is acceptable,
as the only thing that should cause auth failures is misconfiguration of
UNHANGOUT_GOOGLE_CLIENT_ID or UNHANGOUT_GOOGLE_SESSION_SECRET.