Scroll chat only when user hasn't scrolled up
Window will scroll to the bottom only if it is already scrolled to the
bottom, or the user sends a message themselves.  Fixes issue #209.

Tests are incomplete on this feature until we refactor passport mocks to
support multiple users, to test the functionality when a second user
sends a message.