Switch from sass to stylus
Addresses issue #100.

Switch from compass + sass to stylus + nib.  Stylus syntax is almost
identical to SCSS.  The only differences pertinent to what we had is
that it's strict with white-space, and uses " = " for variable
assignment instead of ": ".

This PR uses the connect middleware strategy for compilation that
intercepts all ".css" requests and looks for an associated ".styl" file
to compile as needed.  If we ever got real for-serious traffic such that
static were a bottleneck, we'd want to use a different strategy.