Show admin link to admins on every page
Admin info is stored on the 'event' model -- which makes it tricky to
tell if a user is an admin and should see the "admin" link on any given
page that is not an event page which they administer.

Create a de-duplication of admin-ness on the user model, "adminCache",
which lists the events a user administers.  Use this to determine
whether to show an 'admin' link.