WIP: Front-page require, production-building works
Structure is set out for loading scripts using requirejs, serving
original scripts in development and serving built scripts in production.
Event pages and permalink pages remain unconverted.

For the moment, I've punted on splitting out a "common" set of scripts
which includes vendor-includes such as jquery on every page, distinct
from the page-specific build. Given the low number of distinct page
contexts (front page, admin, event, permalink) it may not be worth the
trouble, and it seems difficult to marry with requirejs-middleware.