Big fat hack to resolve test consistency
Each of the tests passes reliably and consistenly on its own, but some
as-yet unidentified memory issue / race condition / file open problem /
etc. (see issue #269) is causing the test suite to fail half or more of
the time when the entire suite is run at once (but, unfortunately, not
every time).  As a way to resolve this issue, change the `npm test`
definition in `package.json` to execute the test suite in two runs in
two processes: the whole suite minus the most problematic of tests
(`test/test.session-joining.selenium.js`), and the second half with that
problematic portion.

For simplicity, this is done as "all tests starting with `test.[a-r]*`
or `test.[t-z]*`, followed by all tests starting with `test.s*`."  So
far, this causes the test suite to finish each time with high
consistency -- 20 back-to-back runs finished correctly with this setup
on my machine, versus about a 20% success rate with the whole suite
executed at once.  This fix doesn't feel good, but it works.

Also: a couple small code quality fixes.