Remove blur/focus feature
Recent changes to event editing removed the ability to change
"blurEnabled" for events, which bit us hard in a large-scale event.
Additionally, the $(window).blur strategy for detecting blur didn't
properly handle interactions with the youtube video, leading it to have
less utility.

We may consider re-adding it in the future but for now have decided to
remove it, and the concommitant feature for flashing the window title on
new chat messages.