Don't include : if document.location.port is empty
Turns out that IE10 and IE11 report "document.location.port" as empty
string if it's 80 or 443.  That resulted in websocket connection URLs
that look like this:

    wss://unhangout.media.mit.edu:/sock/...
    Note the tricky extra colon -^

On the client, this manifest as a "SyntaxError", presumably arising from
failed parsing of the URL.  Firefox and Chrome had no problem with this
-- it looks like their parsers are just more lax, as they also show
empty string for 80/443.

In testing, this didn't show up because we tend to run test servers on
higher ports (e.g. 7777).

Fixes issue #287.