Move all broadcast logic to backbone listeners
As much as possible, make broadcasting dependent on changes in backbone
models' state, and triggered by events on the models, rather than fired
by other logic.  This is prep-work toward the end of supporting
Operational Transform-style syncing of state.
"lib/unhangout-sockets.js" should be the only server-side place where
Unhangout logic is aware of the fact that there is socket traffic.