Switch to non-insistent video sync indication
Rather than sending an insistent video time sync signal every few
seconds to connected clients, send a signal only on state change and on
first connection.  Modify client to keep track of its own time, and to
gracefully fall out-of-sync if it can't keep up.

Add a time indicator to the "sync" button to hint at what time the
client thinks the server thinks everyone should be at.

NOTE: This is still not adequately tested, due to the high complexity of
attempting to mock YouTube players and the burdensomeness of testing
with actually playing YouTube videos in the testing flow.