Add socket to event-static to reload on open
Add logic to event-static pages which reloads to show the open event if
the event is opened.

Remove the logic on event pages that hides/shows the about pane in
response to the event being opened or closed.  Instead, reload the page
if it is closed and the user isn't an admin.

Update tests to reflect new start/stop conditions.