Restore /admin/ for listing event/session details
The new /myevents/ is awesome, but doesn't display useful detail on the
current state of running sessions and events which the old /admin/
showed.  Restore /admin/ for superusers only, and rename the views,
scripts and styles for /myevents/ under that name.

Also remove some orphaned styles and event handlers for ".scroll-up".