Mock-authenticate while on static URL
To mock-authenticate, we have to get a URL on the site first so that we
can set a domain cookie via ``executeScript``.  We were using the site's
homepage for that, but it's slow, especially with the large Vimeo video
and google spreadsheet now getting pulled down on each pageload.
Instead, always grab the lightest (static) page we have on the site to
do authenticating.

Also increase the selenium startup timeout to 24000 ms (4 minutes) so
that we can have 4 attempts to start the selenium server with
~60000 ms timeouts each before we time out.  Travis is sometimes flakey
about starting selenium.