Fix escaping tests; downgrade firefox
Previously, chat messages were escaped in the model on the server -- a
weird place to do it.  Escaping is now done on render; so move the
testing for that to the browser.

The flexbox we're using for chat input doesn't render properly on FF
24esr, so upgrade firefox -- but not to 31esr, as selenium webdriver
doesn't support that yet.  Pick v28 as the current "goldilocks" which
supports what we need and is supported by selenium.