Fix tricky off-by-factor-of-1000 oplog catchup bug
This is an ugly one, which has likely been part of our world for a
while.

To deal with the space between when a client first requests a page and
first connects a socket, we use an oplog and "catch up" the users for
any traffic they've missed after first connection.  To avoid growing
without bounds, we trim the oplog so that only records the last 60
seconds' ops.  Because of a misplaced seconds/milliseconds conversion,
the oplog was trimming the last 60 milliseconds' events, basically
rendering it ineffective (since 60ms is less than the time it takes to
connect a socket).