Fix title html; fix reload race condition
The presence of an inaccurately hand-escaped closing-p ("&lt;/&gt;"
instead of "&lt;p&gt;") indicated that the hand-escaping was getting
unwieldy, so move popover titles into their own template.

Fix a race condition in browser-reload for NetworkListView. When
reloading the browser, a user could sometimes get its own 'remove
connected user' message as the first message pertaining to its model in
the event, causing the NetworkListView to bind to events on the now
obsolete model, breaking updates for NetworkListView and popovers.
Filter events that we're paying attention to for binding to "add" events
only.