hack open graph call callback in to fix previous video details for now.
This is, to be sure, a pretty ugly hack. It would probably be ok if the
OG stuff was replaced by proper calls to the YouTube v3 Data API, which
I'm not interested in wrestling with right now. The current patch is
meant to at least get things functioning again until somebody can
implement a proper fix.