CLI script for quickly querying and outputting models stored in redis
I got tired of squinting at the single line escaped output that is the stored JSON
of the unhangout redis database. This script provides for a quick and dirty way to
  a) get a list of all keys
  b) get a filtered list of keys
  c) get all data for a specific key
The really nice part is that it's neatly formatted in the CLI output. Run the script
without arguments for usage.