simple event deletion interface.
Adds a delete button to all events on the /admin/ page, with a simple
confirmation dialog. A further server side check is done to ensure
no users are actively connected to the event prior to deletion. The
event and all associated sessions are deleted as part of this action.

Used the node flash module for simple user feedback.