add request destination handling for identity login button.
This exposes the requestDestination parameter to all views, which is
a special querystring parameter than can be appended for redirect
destination handling.

The current implementation leverages the existing destination storage
in the user session related to google auth flow. the _destination
parameter is included in the login link in the _identity view, thus
enabling users to be redirected to the same page they were on after
logging in. while ensureAuthenticated() handles the auth-required
case, this handles the anonymous user case.