fix broken date/time handling on event edit
For events that have a dateAndTime value set, the value of timeZoneValue must
be used to calculate the correct date string for the datetimepicker, otherwise
differences between the admin's local time and the value of the timeZoneValue
setting will result in erroneous formatting of the datetimepicker value.

Using moment-timezone in the view locals allows access to the tz function of
the moment library, while still allowing access to the base moment object.