Add ability to set mock user in dev environment
Allows auth in development environment without access to the Google
APIs (for example, when no internet access is available). Set DEV_AUTH
in conf.json to the Google user ID of the user to authorize as. Note
that the user must already exist in the database for this to work.